refactor: rename AuroraHostListProvider to RdsHostListProvider
