+++ wrapper/src/main/resources/messages.properties
@@ -18,16 +18,16 @@
 AuroraHostListConnectionPlugin.providerAlreadySet=Another dynamic host list provider has already been set: {0}.
 
 # Aurora Host List Provider
-AuroraHostListProvider.clusterInstanceHostPatternRequired=The ''clusterInstanceHostPattern'' configuration property is required when an IP address or custom domain is used to connect to a cluster that provides topology information. If you would instead like to connect without failover functionality, set the 'enableClusterAwareFailover' configuration property to false.
-AuroraHostListProvider.clusterInstanceHostPatternNotSupportedForRDSProxy=An RDS Proxy url can''t be used as the 'clusterInstanceHostPattern' configuration setting.
-AuroraHostListProvider.invalidPattern=Invalid value for the 'clusterInstanceHostPattern' configuration setting - the host pattern must contain a '?' character as a placeholder for the DB instance identifiers of the instances in the cluster.
-AuroraHostListProvider.invalidTopology=The topology query returned an invalid topology - no writer instance detected.
-AuroraHostListProvider.suggestedClusterId=ClusterId ''{0}'' is suggested for url ''{1}''.
-AuroraHostListProvider.parsedListEmpty=Can''t parse connection string: ''{0}''
-AuroraHostListProvider.invalidQuery=Error obtaining host list. Provided database might not be an Aurora Db cluster
-AuroraHostListProvider.errorGettingHostRole=An error occurred while obtaining the connected host's role. This could occur if the connection is broken or if you are not connected to an Aurora database.
-AuroraHostListProvider.errorIdentifyConnection=An error occurred while obtaining the connection's host ID.
-AuroraHostListProvider.errorGettingNetworkTimeout=An error occurred while getting the connection network timeout: {0}
+RdsHostListProvider.clusterInstanceHostPatternRequired=The ''clusterInstanceHostPattern'' configuration property is required when an IP address or custom domain is used to connect to a cluster that provides topology information. If you would instead like to connect without failover functionality, set the 'enableClusterAwareFailover' configuration property to false.
+RdsHostListProvider.clusterInstanceHostPatternNotSupportedForRDSProxy=An RDS Proxy url can''t be used as the 'clusterInstanceHostPattern' configuration setting.
+RdsHostListProvider.invalidPattern=Invalid value for the 'clusterInstanceHostPattern' configuration setting - the host pattern must contain a '?' character as a placeholder for the DB instance identifiers of the instances in the cluster.
+RdsHostListProvider.invalidTopology=The topology query returned an invalid topology - no writer instance detected.
+RdsHostListProvider.suggestedClusterId=ClusterId ''{0}'' is suggested for url ''{1}''.
+RdsHostListProvider.parsedListEmpty=Can''t parse connection string: ''{0}''
+RdsHostListProvider.invalidQuery=Error obtaining host list. Provided database might not be an Aurora Db cluster
+RdsHostListProvider.errorGettingHostRole=An error occurred while obtaining the connected host's role. This could occur if the connection is broken or if you are not connected to an Aurora database.
+RdsHostListProvider.errorIdentifyConnection=An error occurred while obtaining the connection's host ID.
+RdsHostListProvider.errorGettingNetworkTimeout=An error occurred while getting the connection network timeout: {0}
 
 # AWS Credentials Manager
 AwsCredentialsManager.nullProvider=The configured AwsCredentialsProvider was null. If you have configured the AwsCredentialsManager to use a custom AwsCredentialsProviderHandler, please ensure the handler does not return null.
